+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Add whitelist for `socketserver.TCPServer.allow_reuse_address` (Ben Elliston)
 * Clarify that `--exclude` patterns are matched against absolute paths (Jendrik Seipp, #260).
+* Fix example in README file (Jendrik Seipp, #272).
 
 # 2.6 (2022-09-19)